+++ src/3_game_loop/a_game_logic/entities/blight_entity.c
@@ -6,7 +6,7 @@
 /*   By: fschuber <fschuber@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/26 21:57:58 by freddy            #+#    #+#             */
-/*   Updated: 2024/06/27 12:14:12 by fschuber         ###   ########.fr       */
+/*   Updated: 2024/06/27 12:38:15 by fschuber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,17 @@ void	rotate_self(t_entity *self)
 {
 	double	prev_angle;
 	double	new_angle;
+	t_vec2	new_dir;
+
+	if (pos_distance(self->transform.pos, self->spawn_transform.pos) > BLIGHT_MAX_SPAWN_DISTANCE)
+	{
+		new_dir.x = self->spawn_transform.pos.x - self->transform.pos.x;
+		new_dir.y = self->spawn_transform.pos.y - self->transform.pos.y;
+		new_angle = dir_vec_to_deg(new_dir);
+		new_angle += random_val() * BLIGHT_SPAWN_RETURNING_DEG_VARIATION - (BLIGHT_SPAWN_RETURNING_DEG_VARIATION / 2);
+		self->transform.rot = deg_to_dir_vec(new_angle);
+		return ;
+	}
 
 	prev_angle = dir_vec_to_deg(self->transform.rot);
 	self->transform.rot.x = (random_val() * 2) - 1;
@@ -25,14 +36,14 @@ void	rotate_self(t_entity *self)
 		rotate_self(self);
 }
 
-void	tick_blight(t_entity *self)
+void	move_self(t_entity *self)
 {
 	t_vec2	new_pos;
 	t_vec2	oldpos;
 
 	oldpos = self->transform.pos;
 	self->transform.rot = scale_vector(self->transform.rot, 1);
-	if (random_val() < BLIGHT_MOVEMENT_SWITCH_CHANCE)
+	if (random_val() < BLIGHT_MOVEMENT_ROTATION_SWITCH_CHANCE)
 		rotate_self(self);
 	new_pos.x = self->transform.pos.x + self->transform.rot.x * BLIGHT_SPEED;
 	new_pos.y = self->transform.pos.y + self->transform.rot.y * BLIGHT_SPEED;
@@ -47,6 +58,31 @@ void	tick_blight(t_entity *self)
 	{
 		game()->dirty = true;
 		if (MARK_DIRTY_LOGGING)
-			logger(LOGGER_INFO, "Blight moved, markind dirty!");
+			logger(LOGGER_INFO, "Blight moved, marking dirty!");
+	}
+}
+
+void	tick_blight(t_entity *self)
+{
+	t_blight	*blight;
+
+	blight = (t_blight *)self->data;
+	if (blight->state == BLIGHT_STATE_WALKING)
+	{
+		move_self(self);
+		if (random_val() < BLIGHT_STANDING_START_CHANCE)
+		{
+			blight->state = BLIGHT_STATE_STANDING;
+			game()->dirty = true;
+			if (MARK_DIRTY_LOGGING)
+				logger(LOGGER_INFO, "Blight stopped, marking dirty!");
+		}
+	}
+	else if (blight->state == BLIGHT_STATE_STANDING)
+	{
+		if (random_val() < BLIGHT_WALKING_START_CHANCE)
+		{
+			blight->state = BLIGHT_STATE_WALKING;
+		}
 	}
 }